unipdbmap assembly update

+++ unipdbmap.py
@@ -1,5 +1,9 @@
+import os
+import os.path
 import sys
 import Bio
+import time
+import numpy
 import subprocess
 from Bio.PDB.PDBParser import PDBParser
 from Bio.PDB.PDBIO import PDBIO
@@ -9,9 +13,8 @@ from Bio.PDB.PDBIO import Select
 def get_pdbchains(uniprotcode):
     pdblist = []
     chaindic = {}
-    subprocess.run('wget https://www.uniprot.org/uniprot/'+uniprotcode+'.txt', shell=True)
-    try:
-        for line in open(uniprotcode+'.txt', 'r'):
+    if uniprotcode in uniprot_dic:
+        for line in uniprot_dic[uniprotcode]:
  
             if line.startswith('DR   PDB; '):
                 fields = line.split('; ')
@@ -31,13 +34,89 @@ def get_pdbchains(uniprotcode):
                 pdblist.append(pdbcode)
                 chaindic[pdbcode] = [bestchaingroup, maxlength]
 
-        subprocess.run('rm '+uniprotcode+'.txt', shell=True)
         return pdblist, chaindic
 
-    except Exception as e:
-        print(e)
-        subprocess.run('rm '+uniprotcode+'.txt', shell=True)
+    else: 
+        print ("Unavailable data on UniprotKB!")
         return pdblist, chaindic
+
+def derive_assembly(pdbpath):
+
+    rotation = []
+    translation = []
+    matrixgroup = []
+    coordinates = []
+    rototranslations = {}
+    chainlist = ''
+    for line in open(pdbpath, 'r'):
+        chainsrow = False
+        matrixrow = False
+        if line.startswith('REMARK 350'): r350 = True
+        else: r350 = False
+
+        if r350:
+            if line.startswith('REMARK 350 APPLY THE FOLLOWING'): 
+                chainsrow = True
+                chainlist = []
+
+            if chainsrow:
+                line = line.split(': ')[-1].strip().rstrip(',')
+                chainlist += line.split(',')
+                for pos in range(len(chainlist)): chainlist[pos] = chainlist[pos].strip()
+
+            if line.startswith('REMARK 350   BIOMT'):
+                chainsrow = False
+                matrixrow = True
+
+            if matrixrow:
+                model = line.split()[3]
+                rototranslations[model] = rototranslations.get(model, [])
+                rotation.append(line.split()[-4:-1])
+                for pos in range(3): rotation[-1][pos] = float(rotation[-1][pos])
+                translation.append(float(line.split()[-1]))
+
+                if len(rotation) == 3:
+                    rt_matrix = [list(rotation), list(translation)]
+                    added = False
+                    for pos in range(len(rototranslations[model])):
+                        if rt_matrix in rototranslations[model][pos]:
+                            rototranslations[model][pos][0] += chainlist
+                            added = True
+                    if not added: rototranslations[model] += [[chainlist, rt_matrix]]
+                    translation = []
+                    rotation = []
+
+        if line.startswith('ATOM') or line.startswith('TER'): coordinates.append(line)
+
+    rt_structures = []
+
+
+    print (pdbpath)
+    print (rototranslations)
+
+    identity_rtmatrix = [[[1.0, 0.0, 0.0], [0.0, 1.0, 0.0], [0.0, 0.0, 1.0]], [0.0, 0.0, 0.0]]
+
+    outfile = open(pdbpath+'.ass'+str(assembly),'a')
+    for assembly in rototranslations:
+        for rtmatrix in rototranslations[assembly]:
+            if rtmatrix[1] == identity_rtmatrix: continue
+            for chain in rtmatrix[0]:
+                
+#            write_assembly(outfile, rtmatrix, coordinates)
+#
+#
+#def get_rtstructure(outfile, matrix, coordinates):
+#
+#    rotation_matrix = np.array(matrix[1][0])
+#    translation_matrix = np.array(matrix[1][1])
+#
+#    for line in coordinates:
+#        if not line.startswith('TER'):
+#            if line[21] not in matrix[0]: continue
+#            atom_coordinates = np.array([float(line[30:38]), float(line[38:46]), float(line[46:54])])
+#
+#            outfile.write(line[:30])
+        
 
 def extinterface(ichain1, ichain2):
 
@@ -84,9 +163,9 @@ class ChainSelect(Select):
              return False
 
 io=PDBIO()
-p = PDBParser()
-if len(sys.argv) < 4:
-    print ('Usage: python unipdbmap.py [codelist] [outpath] [mode=single/pair/linkedpair]')
+p = PDBParser(QUIET=True)
+if len(sys.argv) < 5:
+    print ('Usage: python unipdbmap.py [codelist] [uniprot_txt_file] [outpath] [mode=single/pair/linkedpair]')
     print ('mode single:')
     print ('---downloads the uniprot id related pdb with the longest available chain.')
     print ('---codelist must contain one uniprot id per line')
@@ -98,8 +177,24 @@ if len(sys.argv) < 4:
     print ('---codelist must contain two uniprot ids for line separated by _')
     sys.exit()
 
-outpath = sys.argv[2]
-mode = sys.argv[3]
+upcode = ''
+uniprot_dic = {'':[]}
+for line in open(sys.argv[2],'r'):
+    if line.startswith("AC   "): 
+        upcode = line[5:-1].split(';')
+        for code in upcode: uniprot_dic[code.strip()] = []
+    else: 
+        for code in upcode: uniprot_dic[code.strip()].append(line.rstrip())
+
+outpath = sys.argv[3]
+mode = sys.argv[4]
+try: pdbdb_path = os.environ['PDBDB']
+except: 
+    print ('WARNING! Local PDB database path not declared!') 
+    print ('To use a local version of the PDB database, declare the path containing all the pdb subfolders (00/ 27/ 51/ .. ) by running:') 
+    print ('export PDBDB=\'local_pdbfolder_path\'')
+    print ('Starting anyway in 5 seconds ...')
+    time.sleep(5)
 
 if mode == 'single':
     for code in open(sys.argv[1],'r'):
@@ -113,16 +208,13 @@ if mode == 'single':
                 longestchain = chaindict[pdb][0]
                 longestpdb = pdb
 
-        subprocess.run('wget https://files.rcsb.org/download/'+longestpdb+'.pdb', shell=True)
+        try: structure = p.get_structure('', pdbdb_path+'/'+longestpdb[1:3].lower()+'/'+'pdb'+longestpdb.lower()+'.ent')
+        except: 
+            subprocess.run('wget https://files.rcsb.org/download/'+longestpdb+'.pdb', shell=True)
+            if os.path.exists(longestpdb+'.pdb'): structure = p.get_structure('', longestpdb+'.pdb')
+            else: continue
 
-        try:
-            structure = p.get_structure('', longestpdb+'.pdb')
-            m, accept = model_chain_select(structure, longestchain[0])
-
-        except Exception as e: 
-            print(e)
-            subprocess.run('rm '+longestpdb+'.pdb', shell=True)
-            continue
+        m, accept = model_chain_select(structure, longestchain[0]) 
 
         io.set_structure(structure)
         io.save(outpath+code.rstrip()+'_'+longestpdb+'_'+accept+'.pdb', 
@@ -155,23 +247,22 @@ if mode == 'pair':
                 longestpdb2 = pdb
 
         if longestpdb1 == '' or longestpdb2 == '': continue
-        subprocess.run('wget https://files.rcsb.org/download/'+longestpdb1+'.pdb',
-                       shell=True)
-        subprocess.run('wget https://files.rcsb.org/download/'+longestpdb2+'.pdb',
-                       shell=True)
 
-        try:
-            structure1 = p.get_structure('', longestpdb1+'.pdb')
-            m1, accept1 = model_chain_select(structure1, longestchain1[0])
-            
-            structure2 = p.get_structure('', longestpdb2+'.pdb')
-            m2, accept2 = model_chain_select(structure2, longestchain2[0])
+        try: structure1 = p.get_structure('', pdbdb_path+'/'+longestpdb1[1:3].lower()+'/'+'pdb'+longestpdb.lower()+'.ent')
+        except: 
+            subprocess.run('wget https://files.rcsb.org/download/'+longestpdb1+'.pdb', shell=True)
+            if os.path.exists(longestpdb1+'.pdb'): structure1 = p.get_structure('', longestpdb1+'.pdb')
+            else: continue
 
-        except Exception as e:
-            print(e)
-            subprocess.run('rm '+longestpdb1+'.pdb', shell=True)
-            subprocess.run('rm '+longestpdb2+'.pdb', shell=True)
-            continue
+        try: structure2 = p.get_structure('', pdbdb_path+'/'+longestpdb2[1:3].lower()+'/'+'pdb'+longestpdb.lower()+'.ent')
+        except: 
+            subprocess.run('wget https://files.rcsb.org/download/'+longestpdb2+'.pdb', shell=True)
+            if os.path.exists(longestpdb2+'.pdb'): structure2 = p.get_structure('', longestpdb2+'.pdb')
+            else: continue
+
+        
+        m1, accept1 = model_chain_select(structure1, longestchain1[0])
+        m2, accept2 = model_chain_select(structure2, longestchain2[0])
 
         lengthchain1 = len(structure1[m1][accept1])
         lengthchain2 = len(structure2[m2][accept2])
@@ -212,7 +303,7 @@ if mode == 'linkedpair':
         code2 = codecouple.split('_')[1].rstrip()
         pdblist1, chaindict1 = get_pdbchains(code1)
         pdblist2, chaindict2 = get_pdbchains(code2)
-    
+ 
         pdblist1=set(pdblist1)
         pdblist2=set(pdblist2)
         commonpdb = list(pdblist1.intersection(pdblist2))
@@ -225,24 +316,31 @@ if mode == 'linkedpair':
                 maxlength = chaindict1[pdb][1]+chaindict2[pdb][1]
                 longestchain = chaindict1[pdb][0]+'-'+chaindict2[pdb][0]
                 longestpdb = pdb
-    
+
+
         if len(longestchain.split('-')) < 2: continue
         chains1 = longestchain.split('-')[0]
         chains2 = longestchain.split('-')[1]
-    
-        subprocess.run('wget https://files.rcsb.org/download/'+longestpdb+'.pdb', 
-                       shell=True)
 
-        try:
-            structure = p.get_structure('', longestpdb+'.pdb')
-            for model in structure:
-                m = model.get_id()
-                break
-            
-            maxlength = 0
-            maxinterface = ''
-            for c1 in chains1.split('/'):
-                for c2 in chains2.split('/'):
+
+        if os.path.exists(pdbdb_path+'/'+longestpdb[1:3].lower()+'/'+'pdb'+longestpdb.lower()+'.ent'):
+            structure = p.get_structure('', pdbdb_path+'/'+longestpdb[1:3].lower()+'/'+'pdb'+longestpdb.lower()+'.ent')
+            reference_path = pdbdb_path+'/'+longestpdb[1:3].lower()+'/'+'pdb'+longestpdb.lower()+'.ent'
+        else:
+            subprocess.run('wget https://files.rcsb.org/download/'+longestpdb+'.pdb', shell=True)
+            if os.path.exists(longestpdb+'.pdb'): 
+                structure = p.get_structure('', longestpdb+'.pdb')
+                reference_path = longestpdb+'.pdb'
+            else: continue
+
+        derive_assembly(reference_path)   
+
+        m = 0
+        maxlength = 0
+        maxinterface = ''
+        for c1 in chains1.split('/'):
+            for c2 in chains2.split('/'):
+                if c1 != c2:
                     int_residues1 = extinterface(structure[m][c1], structure[m][c2])
                     if len(int_residues1) < 5: continue
                     int_residues2 = extinterface(structure[m][c2], structure[m][c1])
@@ -251,16 +349,10 @@ if mode == 'linkedpair':
                     if len(int_residues) > maxlength: 
                         maxlength = len(int_residues)
                         maxinterface = c1+c2
-
-        except Exception as e: 
-            print(e)
-            subprocess.run('rm '+longestpdb+'.pdb', shell=True)
-            continue
     
-        if len(maxinterface)<2: 
-            subprocess.run('rm '+longestpdb+'.pdb', shell=True)
-            continue
+        if len(maxinterface)<2: continue
     
+        print ('Wrinting '+outpath+codecouple.rstrip()+'_'+longestpdb+'_'+maxinterface+'....')
         lengthchain1 = len(structure[m][maxinterface[0]])
         lengthchain2 = len(structure[m][maxinterface[1]])
         if lengthchain1 < lengthchain2: maxinterface = maxinterface[1]+maxinterface[0]
@@ -275,5 +367,5 @@ if mode == 'linkedpair':
         io.save(outpath+codecouple.rstrip()+'_'+longestpdb+'_'+maxinterface+'_u2.pdb', 
                 ChainSelect())
     
-        subprocess.run('rm '+longestpdb+'.pdb', shell=True)
+        if os.path.exists(longestpdb+'.pdb'): subprocess.run('rm '+longestpdb+'.pdb', shell=True)
